BREAKING CHANGE: rename dnsctl to dnsadm
The CLI dnsctl has been renamed to dnsadm. As part of this change
the functionality was moved to a subcommand. So the usage changed
from "dnsctl <domain>" to "dnsadm ddns <domain>". Additionally,
the credentials file can now be overriden with the --auth-file
command line flag and the CREDENTIALS_FILE environment variable
was removed.